support or operator for where conditions (resolves #2)

+++ tests/Query/DeleteQueryTest.php
@@ -33,6 +33,18 @@ class DeleteQueryTest extends PHPUnit_Framework_TestCase
         $this->assertEquals([['balance', '>', 10], ['notes IS NULL']], $where->getConditions());
     }
 
+    public function testOrWhere()
+    {
+        $query = new DeleteQuery();
+
+        $this->assertEquals($query, $query->orWhere('balance', 10, '>'));
+        $this->assertEquals($query, $query->orWhere('notes IS NULL'));
+        $where = $query->getWhere();
+        $this->assertInstanceOf('JAQB\Statement\WhereStatement', $where);
+        $this->assertFalse($where->isHaving());
+        $this->assertEquals([['OR'], ['balance', '>', 10], ['OR'], ['notes IS NULL']], $where->getConditions());
+    }
+
     public function testNot()
     {
         $query = new DeleteQuery();
@@ -93,15 +105,16 @@ class DeleteQueryTest extends PHPUnit_Framework_TestCase
               ->between('created_at', '2016-04-01', '2016-04-30')
               ->notBetween('balance', 100, 150)
               ->not('disabled')
+              ->orWhere('admin', true)
               ->limit(100)
               ->orderBy('uid', 'ASC');
 
         // test for idempotence
         for ($i = 0; $i < 3; ++$i) {
-            $this->assertEquals('DELETE FROM `Users` WHERE `uid` = ? AND `created_at` BETWEEN ? AND ? AND `balance` NOT BETWEEN ? AND ? AND `disabled` <> ? ORDER BY `uid` ASC LIMIT 100', $query->build());
+            $this->assertEquals('DELETE FROM `Users` WHERE `uid` = ? AND `created_at` BETWEEN ? AND ? AND `balance` NOT BETWEEN ? AND ? AND `disabled` <> ? OR `admin` = ? ORDER BY `uid` ASC LIMIT 100', $query->build());
 
             // test values
-            $this->assertEquals([10, '2016-04-01', '2016-04-30', 100, 150, true], $query->getValues());
+            $this->assertEquals([10, '2016-04-01', '2016-04-30', 100, 150, true, true], $query->getValues());
         }
     }
